Factor out extractors. Write Orca extractor.
