+++ stko/calculators/extractors/__init__.py
@@ -1,0 +1,2 @@
+from .xtb_extractor import *  # noqa
+from .orca_extractor import *  # noqa
